Add explicit position tracking for correct reporting
The implicit position tracking with tocIndex was faulty because (1) the index was always one index too far, as ReadNextFrame increases it after reading (it was only correct in the single case that all data of a packet was read without any leftovers in the buffer), and (2) when reading from the beginning of the stream (or always after a seek before the decoder warm-up addition) and the decoder needs an additional frame to return decoded data, the index was even one more too far. By tracking the position explicitly, Position always return the correct position.