Set asio buffer size
On AsioOut.InitRecordAndPlayback i've added a new optional parameter to set the buffer size on the current asio driver instance: if it's not specified (value = -1) it will use the preferred buffer size, else it will check if the value is between min and max size supported and also if it respect granularity; if the value is not coherent with the driver granularity it will search the closest good value and use it.
I've added a public method on AsioOut to get the current buffer size and also added a public property to expose the preferred buffer size (i've founde them to be useful).

I've tried this with different drivers and hardware and it seems to work well.
I've used:
- integrated generic audio card with asio4all
- m-audio m-track plus II with it's own asio driver and also with asio4all
- focusrite scarlet with it's own asio driver and also with asio4all
- behringer u-phoria um2 with it's own asio driver (but now it no longer exists and they force to use asio4all) and also with asio4all
All of this different setups has been tested on 2 different windows computers with windows 10.
Unfortunately i don't have access to any other audio cards or computers at the moment.